chore: move identical methods
Move methods up to consolidate when identical
