+++ hyundai_kia_connect_api/ApiImplType1.py
@@ -2,6 +2,7 @@
 
 import datetime as dt
 from typing import Optional
+import logging
 
 from .ApiImpl import (
     ApiImpl,
@@ -22,6 +23,9 @@ from .const import (
 )
 
 USER_AGENT_OK_HTTP: str = "okhttp/3.12.0"
+
+_LOGGER = logging.getLogger(__name__)
+
 
 
 class ApiImplType1(ApiImpl):
@@ -319,3 +323,125 @@ class ApiImplType1(ApiImpl):
             )
 
         vehicle.data = state
+
+    def update_day_trip_info(
+            self,
+            token,
+            vehicle,
+            yyyymmdd_string,
+        ) -> None:
+            """
+            feature only available for some regions.
+            Updates the vehicle.day_trip_info information for the specified day.
+
+            Default this information is None:
+
+            day_trip_info: DayTripInfo = None
+            """
+            vehicle.day_trip_info = None
+            json_result = self._get_trip_info(
+                token,
+                vehicle,
+                yyyymmdd_string,
+                1,  # day trip info
+            )
+            day_trip_list = json_result["resMsg"]["dayTripList"]
+            if len(day_trip_list) > 0:
+                msg = day_trip_list[0]
+                result = DayTripInfo(
+                    yyyymmdd=yyyymmdd_string,
+                    trip_list=[],
+                    summary=TripInfo(
+                        drive_time=msg["tripDrvTime"],
+                        idle_time=msg["tripIdleTime"],
+                        distance=msg["tripDist"],
+                        avg_speed=msg["tripAvgSpeed"],
+                        max_speed=msg["tripMaxSpeed"],
+                    ),
+                )
+                for trip in msg["tripList"]:
+                    processed_trip = TripInfo(
+                        hhmmss=trip["tripTime"],
+                        drive_time=trip["tripDrvTime"],
+                        idle_time=trip["tripIdleTime"],
+                        distance=trip["tripDist"],
+                        avg_speed=trip["tripAvgSpeed"],
+                        max_speed=trip["tripMaxSpeed"],
+                    )
+                    result.trip_list.append(processed_trip)
+
+                vehicle.day_trip_info = result
+
+                
+    def _get_driving_info(self, token: Token, vehicle: Vehicle) -> dict:
+        url = self.SPA_API_URL + "vehicles/" + vehicle.id + "/drvhistory"
+
+        responseAlltime = requests.post(
+            url,
+            json={"periodTarget": 1},
+            headers=self._get_authenticated_headers(
+                token, vehicle.ccu_ccs2_protocol_support
+            ),
+        )
+        responseAlltime = responseAlltime.json()
+        _LOGGER.debug(f"{DOMAIN} - get_driving_info responseAlltime {responseAlltime}")
+        _check_response_for_errors(responseAlltime)
+
+        response30d = requests.post(
+            url,
+            json={"periodTarget": 0},
+            headers=self._get_authenticated_headers(
+                token, vehicle.ccu_ccs2_protocol_support
+            ),
+        )
+        response30d = response30d.json()
+        _LOGGER.debug(f"{DOMAIN} - get_driving_info response30d {response30d}")
+        _check_response_for_errors(response30d)
+        if get_child_value(responseAlltime, "resMsg.drivingInfo.0"):
+            drivingInfo = responseAlltime["resMsg"]["drivingInfo"][0]
+
+            drivingInfo["dailyStats"] = []
+            if get_child_value(response30d, "resMsg.drivingInfoDetail.0"):
+                for day in response30d["resMsg"]["drivingInfoDetail"]:
+                    processedDay = DailyDrivingStats(
+                        date=dt.datetime.strptime(day["drivingDate"], "%Y%m%d"),
+                        total_consumed=get_child_value(day, "totalPwrCsp"),
+                        engine_consumption=get_child_value(day, "motorPwrCsp"),
+                        climate_consumption=get_child_value(day, "climatePwrCsp"),
+                        onboard_electronics_consumption=get_child_value(
+                            day, "eDPwrCsp"
+                        ),
+                        battery_care_consumption=get_child_value(
+                            day, "batteryMgPwrCsp"
+                        ),
+                        regenerated_energy=get_child_value(day, "regenPwr"),
+                        distance=get_child_value(day, "calculativeOdo"),
+                        distance_unit=vehicle.odometer_unit,
+                    )
+                    drivingInfo["dailyStats"].append(processedDay)
+
+            for drivingInfoItem in response30d["resMsg"]["drivingInfo"]:
+                if (
+                    drivingInfoItem["drivingPeriod"] == 0
+                    and next(
+                        (
+                            v
+                            for k, v in drivingInfoItem.items()
+                            if k.lower() == "calculativeodo"
+                        ),
+                        0,
+                    )
+                    > 0
+                ):
+                    drivingInfo["consumption30d"] = round(
+                        drivingInfoItem["totalPwrCsp"]
+                        / drivingInfoItem["calculativeOdo"]
+                    )
+                    break
+
+            return drivingInfo
+        else:
+            _LOGGER.debug(
+                f"{DOMAIN} - Driving info didn't return valid data. This may be normal if the car doesn't support it."  # noqa
+            )
+            return None
