+++ hyundai_kia_connect_api/KiaUvoApiAU.py
@@ -892,105 +892,6 @@ class KiaUvoApiAU(ApiImplType1):
 
             vehicle.month_trip_info = result
 
-    def update_day_trip_info(
-        self,
-        token,
-        vehicle,
-        yyyymmdd_string,
-    ) -> None:
-        """
-        feature only available for some regions.
-        Updates the vehicle.day_trip_info information for the specified day.
-
-        Default this information is None:
-
-        day_trip_info: DayTripInfo = None
-        """
-        vehicle.day_trip_info = None
-        json_result = self._get_trip_info(
-            token,
-            vehicle,
-            yyyymmdd_string,
-            1,  # day trip info
-        )
-        day_trip_list = json_result["resMsg"]["dayTripList"]
-        if len(day_trip_list) > 0:
-            msg = day_trip_list[0]
-            result = DayTripInfo(
-                yyyymmdd=yyyymmdd_string,
-                trip_list=[],
-                summary=TripInfo(
-                    drive_time=msg["tripDrvTime"],
-                    idle_time=msg["tripIdleTime"],
-                    distance=msg["tripDist"],
-                    avg_speed=msg["tripAvgSpeed"],
-                    max_speed=msg["tripMaxSpeed"],
-                ),
-            )
-            for trip in msg["tripList"]:
-                processed_trip = TripInfo(
-                    hhmmss=trip["tripTime"],
-                    drive_time=trip["tripDrvTime"],
-                    idle_time=trip["tripIdleTime"],
-                    distance=trip["tripDist"],
-                    avg_speed=trip["tripAvgSpeed"],
-                    max_speed=trip["tripMaxSpeed"],
-                )
-                result.trip_list.append(processed_trip)
-            vehicle.day_trip_info = result
-
-    def _get_driving_info(self, token: Token, vehicle: Vehicle) -> dict:
-        url = self.SPA_API_URL + "vehicles/" + vehicle.id + "/drvhistory"
-
-        responseAlltime = requests.post(
-            url,
-            json={"periodTarget": 1},
-            headers=self._get_authenticated_headers(token),
-        )
-        responseAlltime = responseAlltime.json()
-        _LOGGER.debug(f"{DOMAIN} - get_driving_info responseAlltime {responseAlltime}")
-        _check_response_for_errors(responseAlltime)
-
-        response30d = requests.post(
-            url,
-            json={"periodTarget": 0},
-            headers=self._get_authenticated_headers(token),
-        )
-        response30d = response30d.json()
-        _LOGGER.debug(f"{DOMAIN} - get_driving_info response30d {response30d}")
-        _check_response_for_errors(response30d)
-        if get_child_value(responseAlltime, "resMsg.drivingInfoDetail.0"):
-            drivingInfo = responseAlltime["resMsg"]["drivingInfoDetail"][0]
-
-            drivingInfo["dailyStats"] = []
-            for day in response30d["resMsg"]["drivingInfoDetail"]:
-                processedDay = DailyDrivingStats(
-                    date=dt.datetime.strptime(day["drivingDate"], "%Y%m%d"),
-                    total_consumed=day["totalPwrCsp"],
-                    engine_consumption=day["motorPwrCsp"],
-                    climate_consumption=day["climatePwrCsp"],
-                    onboard_electronics_consumption=day["eDPwrCsp"],
-                    battery_care_consumption=day["batteryMgPwrCsp"],
-                    regenerated_energy=day["regenPwr"],
-                    distance=day["calculativeOdo"],
-                    distance_unit=vehicle.odometer_unit,
-                )
-                drivingInfo["dailyStats"].append(processedDay)
-
-            for drivingInfoItem in response30d["resMsg"]["drivingInfo"]:
-                if drivingInfoItem["drivingPeriod"] == 0:
-                    drivingInfo["consumption30d"] = round(
-                        drivingInfoItem["totalPwrCsp"]
-                        / drivingInfoItem["calculativeOdo"]
-                    )
-                    break
-
-            return drivingInfo
-        else:
-            _LOGGER.debug(
-                f"{DOMAIN} - Driving info didn't return valid data. This may be normal if the car doesn't support it."  # noqa
-            )
-            return None
 
     def set_charge_limits(
         self, token: Token, vehicle: Vehicle, ac: int, dc: int
